It Should Have Always Been After: A short commit with a long message!
This does a bytewise search for lines - currently, a search for the
\n before the target line. Mostly, this works.

The problem is that if the last test has "before" on a "\nlow
line\nhigh line", it will read "low line\nhigh line", appropriately
take the line described to be "high line" which is too high, and
so ends with before="\nlow line\nhigh line" (wrong) and after="low
line\nhigh line" (right for this purpose).

Really, just a silly mistake.